Move fixture to conftest, handle db transaction better
